feat: [PQ-210] Steps refactor (#9)
---------
